requirements.txt
#django==1.8
swampdragon==0.4.1.2
swampdragon-auth==0.1.3

# DJANGO Modules
django-model-utils==2.2